Created skeleton for multi-page view
