+++ src/pages/player_gw.py
@@ -1,0 +1,6 @@
+import dash
+from dash import Dash, dcc, html, dash_table, Input, Output
+
+dash.register_page(__name__, name="Player GW Statistics")
+
+layout = None
